[api] Adding support for Dialog Queries
- Primary /api/service_dialogs collection queries
- Primary /api/provision_dialogs collection queries
- Making DialogYamlSerializer a generic DialogSerializer
- Updating DialogYamlSerializer to subclass DialogSerializer
- Adding content virtual attribute to Dialog to mimic provision dialogs
- Adding virtual reflection :dialogs to service_template and service
- Adding service_dialogs.rb to fetch the dialog content when expanding resource
- Subcollection /api/service_templates/:id/service_dialogs
- Subcollection /api/services/:id/service_dialogs

https://trello.com/c/88TGJWqP
