+++ app/controllers/api_controller.rb
@@ -65,6 +65,7 @@ class ApiController < ApplicationController
   include_concern 'RequestTasks'
   include_concern 'Roles'
   include_concern 'ServiceCatalogs'
+  include_concern 'ServiceDialogs'
   include_concern 'ServiceRequests'
   include_concern 'Software'
   include_concern 'ServiceTemplates'
